+++ GetXYRatio.h
@@ -1,6 +1,6 @@
 #ifndef GETXYRATIO_H
 #define GETXYRATIO_H
 
-extern void GetXYRatio(int *xRatio, int *yRatio, int x, int y, int PlayerX, int PlayerY);
+extern void GetXYRatio(double *xRatio, double *yRatio, double x, double y, double PlayerX, double PlayerY);
 
 #endif
